requirements.txt
Django==5.1.4
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.4.0
psycopg[binary]==3.2.1
python-dotenv==1.0.1
requests==2.32.3
gunicorn==21.2.0
dj-database-url==2.1.0
whitenoise==6.6.0
